+++ src/singleTable/implementation.ts
@@ -159,14 +159,14 @@ export class SingleTable<SingleParams extends SingleTableParams> {
     return this.methods.executeTransaction(configs);
   }
 
-  generateTransactionConfigList<Item>(
+  generateTransactionConfigList<Item extends AnyObject>(
     items: Item[],
     generator: SingleTableTransactConfigGenerator<Item, SingleParams>,
-  ): SingleTableTransactionConfig<SingleParams>[] {
+  ): SingleTableTransactionConfig<SingleParams, Item>[] {
     return this.methods.generateTransactionConfigList(
       items,
       generator,
-    ) as SingleTableTransactionConfig<SingleParams>[];
+    ) as SingleTableTransactionConfig<SingleParams, Item>[];
   }
 
   createSet(items: string[]): DBSet {