+++ changelog.md
@@ -1,5 +1,35 @@
 # DynamoDB Provider Changelog
 
+## v0.0.6
+
+- `transact` params generation added to entity. Now you can simplify transact params creations:
+
+```ts
+await table.executeTransaction([
+  {
+    create: entity1.getCreationParams({...})
+  },
+  {
+    create: entity2.getCreationParams({...})
+  },
+  // ...
+])
+
+// to
+
+await table.executeTransaction([
+  entity1.transactCreateParams({...}),
+
+  entity2.transactUpdateParams({...}),
+
+  entity3.transactDeleteParams({...}),
+
+  entity4.transactValidadeParams({...}),
+  // ...
+])
+```
+
+
 ## v0.0.5
 
 - `blockInternalPropUpdate` configuration added to SingleTable. now by default, any update that references any internal table prop will throw. You can disable this behavior and use the `badUpdateValidation` to only block the ones you want